create service specific tinylicious fluid static
